fix: check enable_social_features before flushing followers cache

+++ application/helpers/importer.helper.php
@@ -237,8 +237,10 @@ $importer->finish = function() use ($redis) {
     $redis->delete("tags_user_{$this->user->id}_public");
     $redis->delete("tags_user_{$this->user->id}_private");
     # Flush Friends Feeds
-    foreach ($this->user->follower_ids as $user_id) {
-      $redis->delete("feed_marks_friends_{$user_id}}");
+    if (flag('enable_social_features')) {
+      foreach ($this->user->follower_ids as $user_id) {
+        $redis->delete("feed_marks_friends_{$user_id}}");
+      }
     }
   }
 };
